requirements.txt
absl-py
aiosignal
attrs
blinker
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
colorama
contourpy
cryptography
cycler
dlib
face-recognition
face-recognition-models
filelock
firebase-admin
Flask
flatbuffers
fonttools
fpdf
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpcio
grpcio-status
h5py
httplib2
idna
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
kiwisolver
MarkupSafe
matplotlib
mediapipe
msgpack
neovim
numpy
opencv-contrib-python
packaging
pandas
Pillow
proto-plus
protobuf
pyasn1
pyasn1-modules
pycparser
PyJWT
pyparsing
python-dateutil
pytz
PyYAML
ray
referencing
requests
rpds-py
rsa
six
sounddevice
tabulate
typing_extensions
tzdata
uritemplate
urllib3
Werkzeug